Struct google_dfareporting2d3::Ad

pub struct Ad {
    pub click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>,
    pub campaign_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub geo_targeting: Option<GeoTargeting>,
    pub comments: Option<String>,
    pub delivery_schedule: Option<DeliverySchedule>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub event_tag_overrides: Option<Vec<EventTagOverride>>,
    pub id: Option<String>,
    pub size: Option<Size>,
    pub archived: Option<bool>,
    pub technology_targeting: Option<TechnologyTargeting>,
    pub creative_rotation: Option<CreativeRotation>,
    pub click_through_url: Option<ClickThroughUrl>,
    pub creative_group_assignments: Option<Vec<CreativeGroupAssignment>>,
    pub remarketing_list_expression: Option<ListTargetingExpression>,
    pub id_dimension_value: Option<DimensionValue>,
    pub account_id: Option<String>,
    pub type_: Option<String>,
    pub ssl_required: Option<bool>,
    pub audience_segment_id: Option<String>,
    pub key_value_targeting_expression: Option<KeyValueTargetingExpression>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>,
    pub create_info: Option<LastModifiedInfo>,
    pub start_time: Option<String>,
    pub active: Option<bool>,
    pub compatibility: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub day_part_targeting: Option<DayPartTargeting>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub ssl_compliant: Option<bool>,
    pub end_time: Option<String>,
    pub dynamic_click_tracker: Option<bool>,
    pub placement_assignments: Option<Vec<PlacementAssignment>>,
}

Contains properties of a DCM ad.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch ads (request|response)

list ads (none)

insert ads (request|response)

get ads (response)

update ads (request|response)

Fields
click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>

Click-through URL suffix properties for this ad. Applies to the URL in the ad or (if overriding ad properties) the URL in the creative.

campaign_id: Option<String>

Campaign ID of this ad. This is a required field on insertion.

advertiser_id: Option<String>

Advertiser ID of this ad. This is a required field on insertion.

geo_targeting: Option<GeoTargeting>

Geographical targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

comments: Option<String>

Comments for this ad.

delivery_schedule: Option<DeliverySchedule>

Delivery schedule information for this ad. Applicable when type is AD_SERVING_STANDARD_AD or AD_SERVING_TRACKING. This field along with subfields priority and impressionRatio are required on insertion when type is AD_SERVING_STANDARD_AD.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

event_tag_overrides: Option<Vec<EventTagOverride>>

Event tag overrides for this ad.

id: Option<String>

ID of this ad. This is a read-only, auto-generated field.

size: Option<Size>

Size of this ad. Applicable when type is AD_SERVING_DEFAULT_AD.

archived: Option<bool>

Whether this ad is archived.

technology_targeting: Option<TechnologyTargeting>

Technology platform targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

creative_rotation: Option<CreativeRotation>

Creative rotation for this ad. Applicable when type is AD_SERVING_DEFAULT_AD, AD_SERVING_STANDARD_AD, or AD_SERVING_TRACKING. When type is AD_SERVING_DEFAULT_AD, this field should have exactly one creativeAssignment.

click_through_url: Option<ClickThroughUrl>

Click-through URL for this ad. This is a required field on insertion. Applicable when type is AD_SERVING_CLICK_TRACKER.

creative_group_assignments: Option<Vec<CreativeGroupAssignment>>

Creative group assignments for this ad. Applicable when type is AD_SERVING_CLICK_TRACKER. Only one assignment per creative group number is allowed for a maximum of two assignments.

remarketing_list_expression: Option<ListTargetingExpression>

Remarketing list targeting expression for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this ad. This is a read-only, auto-generated field.

account_id: Option<String>

Account ID of this ad. This is a read-only field that can be left blank.

type_: Option<String>

Type of ad. This is a required field on insertion. Note that default ads (AD_SERVING_DEFAULT_AD) cannot be created directly (see Creative resource).

ssl_required: Option<bool>

Whether this ad requires ssl. This is a read-only field that is auto-generated when the ad is inserted or updated.

audience_segment_id: Option<String>

Audience segment ID that is being targeted for this ad. Applicable when type is AD_SERVING_STANDARD_AD.

key_value_targeting_expression: Option<KeyValueTargetingExpression>

Key-value targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

campaign_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>

Default click-through event tag properties for this ad.

create_info: Option<LastModifiedInfo>

Information about the creation of this ad. This is a read-only field.

start_time: Option<String>

Date and time that this ad should start serving. If creating an ad, this field must be a time in the future. This is a required field on insertion.

active: Option<bool>

Whether this ad is active.

compatibility: Option<String>

Compatibility of this ad. Applicable when type is AD_SERVING_DEFAULT_AD. WEB and WEB_INTERSTITIAL refer to rendering either on desktop or on mobile devices for regular or interstitial ads, respectively. APP and APP_INTERSTITIAL are for rendering in mobile apps. IN_STREAM_VIDEO refers to rendering an in-stream video ads developed with the VAST standard.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#ad".

subaccount_id: Option<String>

Subaccount ID of this ad. This is a read-only field that can be left blank.

name: Option<String>

Name of this ad. This is a required field and must be less than 256 characters long.

day_part_targeting: Option<DayPartTargeting>

Time and day targeting information for this ad. This field must be left blank if the ad is using a targeting template. Applicable when type is AD_SERVING_STANDARD_AD.

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this ad. This is a read-only field.

ssl_compliant: Option<bool>

Whether this ad is ssl compliant. This is a read-only field that is auto-generated when the ad is inserted or updated.

end_time: Option<String>

Date and time that this ad should stop serving. Must be later than the start time. This is a required field on insertion.

dynamic_click_tracker: Option<bool>

Whether this ad is a dynamic click tracker. Applicable when type is AD_SERVING_CLICK_TRACKER. This is a required field on insert, and is read-only after insert.

placement_assignments: Option<Vec<PlacementAssignment>>

Placement assignments for this ad.