Struct google_dfareporting2d3::CampaignCreativeAssociationMethods

pub struct CampaignCreativeAssociationMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on campaignCreativeAssociation resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d3 as dfareporting2d3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d3::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `insert(...)` and `list(...)`
// to build up your call.
let rb = hub.campaign_creative_associations();

Methods

impl<'a, C, A> CampaignCreativeAssociationMethods<'a, C, A>

fn insert(&self, request: CampaignCreativeAssociation, profile_id: &str, campaign_id: &str) -> CampaignCreativeAssociationInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Associates a creative with the specified campaign. This method creates a default ad with dimensions matching the creative in the campaign if such a default ad does not exist already.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

campaignId - Campaign ID in this association.

fn list(&self, profile_id: &str, campaign_id: &str) -> CampaignCreativeAssociationListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of creative IDs associated with the specified campaign.

Arguments

profileId - User profile ID associated with this request.

campaignId - Campaign ID in this association.