Struct google_dfareporting2d3::CreativeAsset

pub struct CreativeAsset {
    pub mime_type: Option<String>,
    pub artwork_type: Option<String>,
    pub position_left_unit: Option<String>,
    pub collapsed_size: Option<Size>,
    pub flash_version: Option<i32>,
    pub pushdown: Option<bool>,
    pub hide_flash_objects: Option<bool>,
    pub detected_features: Option<Vec<String>>,
    pub duration: Option<i32>,
    pub position_top_unit: Option<String>,
    pub id: Option<String>,
    pub alignment: Option<String>,
    pub size: Option<Size>,
    pub horizontally_locked: Option<bool>,
    pub start_time_type: Option<String>,
    pub display_type: Option<String>,
    pub expanded_dimension: Option<Size>,
    pub vertically_locked: Option<bool>,
    pub role: Option<String>,
    pub zip_filesize: Option<String>,
    pub bit_rate: Option<i32>,
    pub window_mode: Option<String>,
    pub progressive_serving_url: Option<String>,
    pub duration_type: Option<String>,
    pub video_duration: Option<f32>,
    pub file_size: Option<String>,
    pub offset: Option<OffsetPosition>,
    pub active: Option<bool>,
    pub original_backup: Option<bool>,
    pub hide_selection_boxes: Option<bool>,
    pub streaming_serving_url: Option<String>,
    pub z_index: Option<i32>,
    pub child_asset_type: Option<String>,
    pub backup_image_exit: Option<CreativeCustomEvent>,
    pub action_script3: Option<bool>,
    pub pushdown_duration: Option<f32>,
    pub asset_identifier: Option<CreativeAssetId>,
    pub transparency: Option<bool>,
    pub ssl_compliant: Option<bool>,
    pub position: Option<OffsetPosition>,
    pub custom_start_time_value: Option<i32>,
    pub zip_filename: Option<String>,
}

Creative Asset.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creative assets (none)

Fields
mime_type: Option<String>

Detected MIME type for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

artwork_type: Option<String>

Artwork type of rich media creative. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

position_left_unit: Option<String>

Offset left unit for an asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

collapsed_size: Option<Size>

Size of an asset when collapsed. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA and all VPAID. Additionally, applicable to assets whose displayType is ASSET_DISPLAY_TYPE_EXPANDING or ASSET_DISPLAY_TYPE_PEEL_DOWN.

flash_version: Option<i32>

Flash version of the asset. This is a read-only field. Applicable to the following creative types: FLASH_INPAGE, all RICH_MEDIA, and all VPAID. Applicable to ENHANCED_BANNER when the primary asset type is not HTML_IMAGE.

pushdown: Option<bool>

Whether the asset pushes down other content. Applicable to the following creative types: all RICH_MEDIA. Additionally, only applicable when the asset offsets are 0, the collapsedSize.width matches size.width, and the collapsedSize.height is less than size.height.

hide_flash_objects: Option<bool>

Whether to hide Flash objects flag for an asset. Applicable to the following creative types: all RICH_MEDIA.

detected_features: Option<Vec<String>>

List of feature dependencies for the creative asset that are detected by DCM. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative correctly. This is a read-only, auto-generated field. Applicable to the following creative types: ENHANCED_BANNER and HTML5_BANNER.

duration: Option<i32>

Duration in seconds for which an asset will be displayed. Applicable to the following creative types: INSTREAM_VIDEO and VPAID_LINEAR.

position_top_unit: Option<String>

Offset top unit for an asset. This is a read-only field if the asset displayType is ASSET_DISPLAY_TYPE_OVERLAY. Applicable to the following creative types: all RICH_MEDIA.

id: Option<String>

Numeric ID of this creative asset. This is a required field and should not be modified. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

alignment: Option<String>

Possible alignments for an asset. This is a read-only field. Applicable to the following creative types: RICH_MEDIA_MULTI_FLOATING.

size: Option<Size>

Size associated with this creative asset. This is a required field when applicable; however for IMAGE and FLASH_INPAGE creatives, if left blank, this field will be automatically set using the actual size of the associated image asset. Applicable to the following creative types: ENHANCED_IMAGE, FLASH_INPAGE, HTML5_BANNER, IMAGE, and all RICH_MEDIA. Applicable to ENHANCED_BANNER when the primary asset type is not HTML_IMAGE.

horizontally_locked: Option<bool>

Whether the asset is horizontally locked. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

start_time_type: Option<String>

Initial wait time type before making the asset visible. Applicable to the following creative types: all RICH_MEDIA.

display_type: Option<String>

Type of rich media asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

expanded_dimension: Option<Size>

Detected expanded dimension for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

vertically_locked: Option<bool>

Whether the asset is vertically locked. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

role: Option<String>

Role of the asset in relation to creative. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT. This is a required field. PRIMARY applies to ENHANCED_BANNER, FLASH_INPAGE, HTML5_BANNER, IMAGE, IMAGE_GALLERY, all RICH_MEDIA (which may contain multiple primary assets), and all VPAID creatives. BACKUP_IMAGE applies to ENHANCED_BANNER, FLASH_INPAGE, HTML5_BANNER, all RICH_MEDIA, and all VPAID creatives. ADDITIONAL_IMAGE and ADDITIONAL_FLASH apply to FLASH_INPAGE creatives. OTHER refers to assets from sources other than DCM, such as Studio uploaded assets, applicable to all RICH_MEDIA and all VPAID creatives. PARENT_VIDEO refers to videos uploaded by the user in DCM and is applicable to INSTREAM_VIDEO and VPAID_LINEAR creatives. TRANSCODED_VIDEO refers to videos transcoded by DCM from PARENT_VIDEO assets and is applicable to INSTREAM_VIDEO and VPAID_LINEAR creatives. ALTERNATE_VIDEO refers to the DCM representation of child asset videos from Studio, and is applicable to VPAID_LINEAR_VIDEO creatives. These cannot be added or removed within DCM. For VPAID_LINEAR creatives, PARENT_VIDEO, TRANSCODED_VIDEO and ALTERNATE_VIDEO assets that are marked active serve as backup in case the VPAID creative cannot be served. Only PARENT_VIDEO assets can be added or removed for an INSTREAM_VIDEO or VPAID_LINEAR creative.

zip_filesize: Option<String>

Size of zip file. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.

bit_rate: Option<i32>

Detected bit-rate for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

window_mode: Option<String>

Window mode options for flash assets. Applicable to the following creative types: FLASH_INPAGE, RICH_MEDIA_EXPANDING, RICH_MEDIA_IM_EXPAND, RICH_MEDIA_INPAGE, and RICH_MEDIA_INPAGE_FLOATING.

progressive_serving_url: Option<String>

Progressive URL for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

duration_type: Option<String>

Duration type for which an asset will be displayed. Applicable to the following creative types: all RICH_MEDIA.

video_duration: Option<f32>

Detected video duration for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

file_size: Option<String>

File size associated with this creative asset. This is a read-only field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

offset: Option<OffsetPosition>

Offset position for an asset in collapsed mode. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA and all VPAID. Additionally, only applicable to assets whose displayType is ASSET_DISPLAY_TYPE_EXPANDING or ASSET_DISPLAY_TYPE_PEEL_DOWN.

active: Option<bool>

Whether the video asset is active. This is a read-only field for VPAID_NON_LINEAR assets. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

original_backup: Option<bool>

Whether the backup asset is original or changed by the user in DCM. Applicable to the following creative types: all RICH_MEDIA.

hide_selection_boxes: Option<bool>

Whether to hide selection boxes flag for an asset. Applicable to the following creative types: all RICH_MEDIA.

streaming_serving_url: Option<String>

Streaming URL for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

z_index: Option<i32>

zIndex value of an asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.Additionally, only applicable to assets whose displayType is NOT one of the following types: ASSET_DISPLAY_TYPE_INPAGE or ASSET_DISPLAY_TYPE_OVERLAY.

child_asset_type: Option<String>

Rich media child asset type. This is a read-only field. Applicable to the following creative types: all VPAID.

backup_image_exit: Option<CreativeCustomEvent>

Exit event configured for the backup image. Applicable to the following creative types: all RICH_MEDIA.

action_script3: Option<bool>

Whether ActionScript3 is enabled for the flash asset. This is a read-only field. Applicable to the following creative type: FLASH_INPAGE. Applicable to ENHANCED_BANNER when the primary asset type is not HTML_IMAGE.

pushdown_duration: Option<f32>

Pushdown duration in seconds for an asset. Must be between 0 and 9.99. Applicable to the following creative types: all RICH_MEDIA.Additionally, only applicable when the asset pushdown field is true, the offsets are 0, the collapsedSize.width matches size.width, and the collapsedSize.height is less than size.height.

asset_identifier: Option<CreativeAssetId>

Identifier of this asset. This is the same identifier returned during creative asset insert operation. This is a required field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

transparency: Option<bool>

Whether the asset is transparent. Applicable to the following creative types: all RICH_MEDIA. Additionally, only applicable to HTML5 assets.

ssl_compliant: Option<bool>

Whether the asset is SSL-compliant. This is a read-only field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

position: Option<OffsetPosition>

Offset position for an asset. Applicable to the following creative types: all RICH_MEDIA.

custom_start_time_value: Option<i32>

Custom start time in seconds for making the asset visible. Applicable to the following creative types: all RICH_MEDIA.

zip_filename: Option<String>

File name of zip file. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.