Struct google_dfareporting2d3::FloodlightActivity

pub struct FloodlightActivity {
    pub secure: Option<bool>,
    pub id_dimension_value: Option<DimensionValue>,
    pub floodlight_activity_group_name: Option<String>,
    pub image_tag_enabled: Option<bool>,
    pub advertiser_id: Option<String>,
    pub cache_busting_type: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub floodlight_activity_group_id: Option<String>,
    pub user_defined_variable_types: Option<Vec<String>>,
    pub tag_string: Option<String>,
    pub floodlight_activity_group_tag_string: Option<String>,
    pub expected_url: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub publisher_tags: Option<Vec<FloodlightActivityPublisherDynamicTag>>,
    pub ssl_compliant: Option<bool>,
    pub notes: Option<String>,
    pub floodlight_configuration_id: Option<String>,
    pub id: Option<String>,
    pub counting_method: Option<String>,
    pub default_tags: Option<Vec<FloodlightActivityDynamicTag>>,
    pub floodlight_configuration_id_dimension_value: Option<DimensionValue>,
    pub tag_format: Option<String>,
    pub hidden: Option<bool>,
    pub account_id: Option<String>,
    pub ssl_required: Option<bool>,
    pub floodlight_activity_group_type: Option<String>,
}

Contains properties of a Floodlight activity.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get floodlight activities (response)

insert floodlight activities (request|response)

patch floodlight activities (request|response)

update floodlight activities (request|response)

Fields

secure

Whether this tag should use SSL.

id_dimension_value

Dimension value for the ID of this floodlight activity. This is a read-only, auto-generated field.

floodlight_activity_group_name

Name of the associated floodlight activity group. This is a read-only field.

image_tag_enabled

Whether the image tag is enabled for this activity.

advertiser_id

Advertiser ID of this floodlight activity. If this field is left blank, the value will be copied over either from the activity group's advertiser or the existing activity's advertiser.

cache_busting_type

Code type used for cache busting in the generated tag.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

floodlight_activity_group_id

Floodlight activity group ID of this floodlight activity. This is a required field.

user_defined_variable_types

List of the user-defined variables used by this conversion tag. These map to the "u[1-20]=" in the tags. Each of these can have a user defined type. Acceptable values are: - "U1" - "U2" - "U3" - "U4" - "U5" - "U6" - "U7" - "U8" - "U9" - "U10" - "U11" - "U12" - "U13" - "U14" - "U15" - "U16" - "U17" - "U18" - "U19" - "U20"

tag_string

Value of the cat= paramter in the floodlight tag, which the ad servers use to identify the activity. This is optional: if empty, a new tag string will be generated for you. This string must be 1 to 8 characters long, with valid characters being [a-z][A-Z][0-9][-][ _ ]. This tag string must also be unique among activities of the same activity group. This field is read-only after insertion.

floodlight_activity_group_tag_string

Tag string of the associated floodlight activity group. This is a read-only field.

expected_url

URL where this tag will be deployed. If specified, must be less than 256 characters long.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#floodlightActivity".

subaccount_id

Subaccount ID of this floodlight activity. This is a read-only field that can be left blank.

name

Name of this floodlight activity. This is a required field. Must be less than 129 characters long and cannot contain quotes.

publisher_tags

Publisher dynamic floodlight tags.

ssl_compliant

Whether the floodlight activity is SSL-compliant. This is a read-only field, its value detected by the system from the floodlight tags.

notes

General notes or implementation instructions for the tag.

floodlight_configuration_id

Floodlight configuration ID of this floodlight activity. If this field is left blank, the value will be copied over either from the activity group's floodlight configuration or from the existing activity's floodlight configuration.

id

ID of this floodlight activity. This is a read-only, auto-generated field.

counting_method

Counting method for conversions for this floodlight activity. This is a required field.

default_tags

Dynamic floodlight tags.

floodlight_configuration_id_dimension_value

Dimension value for the ID of the floodlight configuration. This is a read-only, auto-generated field.

tag_format

Tag format type for the floodlight activity. If left blank, the tag format will default to HTML.

hidden

Whether this activity is archived.

account_id

Account ID of this floodlight activity. This is a read-only field that can be left blank.

ssl_required

Whether this floodlight activity must be SSL-compliant.

floodlight_activity_group_type

Type of the associated floodlight activity group. This is a read-only field.