Struct google_dfareporting2d3::PricingSchedulePricingPeriod

pub struct PricingSchedulePricingPeriod {
    pub units: Option<String>,
    pub start_date: Option<String>,
    pub rate_or_cost_nanos: Option<String>,
    pub end_date: Option<String>,
    pub pricing_comment: Option<String>,
}

Pricing Period

This type is not used in any activity, and only used as part of another schema.

Fields

units

Units of this pricing period.

start_date

Pricing period start date. This date must be later than, or the same day as, the placement start date. The hours, minutes, and seconds of the start date should not be set, as doing so will result in an error.

rate_or_cost_nanos

Rate or cost of this pricing period.

end_date

Pricing period end date. This date must be later than, or the same day as, the pricing period start date, but not later than the placement end date. The period end date can be the same date as the period start date. If, for example, you set 6/25/2015 as both the start and end dates, the effective pricing period date is just that day only, 6/25/2015. The hours, minutes, and seconds of the end date should not be set, as doing so will result in an error.

pricing_comment

Comments for this pricing period.