Struct google_dfareporting2d3::ReportCrossDimensionReachCriteria

pub struct ReportCrossDimensionReachCriteria {
    pub breakdown: Option<Vec<SortedDimension>>,
    pub pivoted: Option<bool>,
    pub overlap_metric_names: Option<Vec<String>>,
    pub metric_names: Option<Vec<String>>,
    pub date_range: Option<DateRange>,
    pub dimension_filters: Option<Vec<DimensionValue>>,
    pub dimension: Option<String>,
}

The report criteria for a report of type "CROSS_DIMENSION_REACH".

This type is not used in any activity, and only used as part of another schema.

Fields

breakdown

The list of dimensions the report should include.

pivoted

Whether the report is pivoted or not. Defaults to true.

overlap_metric_names

The list of names of overlap metrics the report should include.

metric_names

The list of names of metrics the report should include.

date_range

The date range this report should be run for.

dimension_filters

The list of filters on which dimensions are filtered.

dimension

The dimension option.