Struct google_dfareporting2d3::ReportSchedule

pub struct ReportSchedule {
    pub start_date: Option<String>,
    pub runs_on_day_of_month: Option<String>,
    pub expiration_date: Option<String>,
    pub active: Option<bool>,
    pub every: Option<i32>,
    pub repeats: Option<String>,
    pub repeats_on_week_days: Option<Vec<String>>,
}

The report's schedule. Can only be set if the report's 'dateRange' is a relative date range and the relative date range is not "TODAY".

This type is not used in any activity, and only used as part of another schema.

Fields
start_date: Option<String>

Start date of date range for which scheduled reports should be run.

runs_on_day_of_month: Option<String>

Enum to define for "MONTHLY" scheduled reports whether reports should be repeated on the same day of the month as "startDate" or the same day of the week of the month. Example: If 'startDate' is Monday, April 2nd 2012 (2012-04-02), "DAY_OF_MONTH" would run subsequent reports on the 2nd of every Month, and "WEEK_OF_MONTH" would run subsequent reports on the first Monday of the month.

expiration_date: Option<String>

The expiration date when the scheduled report stops running.

active: Option<bool>

Whether the schedule is active or not. Must be set to either true or false.

every: Option<i32>

Defines every how many days, weeks or months the report should be run. Needs to be set when "repeats" is either "DAILY", "WEEKLY" or "MONTHLY".

repeats: Option<String>

The interval for which the report is repeated. Note: - "DAILY" also requires field "every" to be set. - "WEEKLY" also requires fields "every" and "repeatsOnWeekDays" to be set. - "MONTHLY" also requires fields "every" and "runsOnDayOfMonth" to be set.

repeats_on_week_days: Option<Vec<String>>

List of week days "WEEKLY" on which scheduled reports should run.