Struct google_dfareporting2d3::Site

pub struct Site {
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub directory_site_id_dimension_value: Option<DimensionValue>,
    pub site_settings: Option<SiteSettings>,
    pub id: Option<String>,
    pub directory_site_id: Option<String>,
    pub site_contacts: Option<Vec<SiteContact>>,
    pub key_name: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub approved: Option<bool>,
    pub account_id: Option<String>,
}

Contains properties of a site.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update sites (request|response)

list sites (none)

insert sites (request|response)

get sites (response)

patch sites (request|response)

Fields
kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#site".

subaccount_id: Option<String>

Subaccount ID of this site. This is a read-only field that can be left blank.

name: Option<String>

Name of this site.This is a required field. Must be less than 128 characters long. If this site is under a subaccount, the name must be unique among sites of the same subaccount. Otherwise, this site is a top-level site, and the name must be unique among top-level sites of the same account.

directory_site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the directory site. This is a read-only, auto-generated field.

site_settings: Option<SiteSettings>

Site-wide settings.

id: Option<String>

ID of this site. This is a read-only, auto-generated field.

directory_site_id: Option<String>

Directory site associated with this site. This is a required field that is read-only after insertion.

site_contacts: Option<Vec<SiteContact>>

Site contacts.

key_name: Option<String>

Key name of this site. This is a read-only, auto-generated field.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this site. This is a read-only, auto-generated field.

approved: Option<bool>

Whether this site is approved.

account_id: Option<String>

Account ID of this site. This is a read-only field that can be left blank.