Struct google_dfareporting2d3::TargetableRemarketingList

pub struct TargetableRemarketingList {
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub description: Option<String>,
    pub list_size: Option<String>,
    pub life_span: Option<String>,
    pub advertiser_id: Option<String>,
    pub list_source: Option<String>,
    pub active: Option<bool>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub account_id: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
}

Contains properties of a targetable remarketing list. Remarketing enables you to create lists of users who have performed specific actions on a site, then target ads to members of those lists. This resource is a read-only view of a remarketing list to be used to faciliate targeting ads to specific lists. Remarketing lists that are owned by your advertisers and those that are shared to your advertisers or account are accessible via this resource. To manage remarketing lists that are owned by your advertisers, use the RemarketingLists resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list targetable remarketing lists (none)

get targetable remarketing lists (response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#targetableRemarketingList".

subaccount_id

Subaccount ID of this remarketing list. This is a read-only, auto-generated field that is only returned in GET requests.

description

Targetable remarketing list description.

list_size

Number of users currently in the list. This is a read-only field.

life_span

Number of days that a user should remain in the targetable remarketing list without an impression.

advertiser_id

Dimension value for the advertiser ID that owns this targetable remarketing list.

list_source

Product from which this targetable remarketing list was originated.

active

Whether this targetable remarketing list is active.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser.

account_id

Account ID of this remarketing list. This is a read-only, auto-generated field that is only returned in GET requests.

id

Targetable remarketing list ID.

name

Name of the targetable remarketing list. Is no greater than 128 characters long.